src/main/python/server/requirements.txt
pytest
pytest-cov
pytest-html
flask
flask-restful
flask-cors
werkzeug>=0.14
argon2_cffi
flask_httpauth
uwsgi